Moves types to their own files
